+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Optimized the `internal/stack.FunctionID` for decreased allocations and CPU usage
+
 ## v3.135.12
 * Optimized the `internal/meta.TraceID` generation of `x-ydb-trace-id` gRPC header